refactor: rename Form/data-context to Form/DataContext for PascalCase consistency
Rename the data-context directory to DataContext to follow the PascalCase
naming convention used by all other directories in the Form/ directory
(e.g., Isolation, Handler, FieldProps). This is an internal directory
rename with no public API changes.

Updated imports in:
- Form/index.ts
- hooks/useSnapshot.tsx
- Form/Isolation/__tests__/Isolation.test.tsx
