+++ packages/dnb-eufemia/src/extensions/forms/Form/index.ts
@@ -13,11 +13,11 @@ export { default as InfoOverlay } from './InfoOverlay'
 export { default as Card } from './Card'
 export { default as Isolation } from './Isolation'
 export { default as Snapshot } from './Snapshot'
-export { default as useData } from './data-context/useData'
-export { default as setData } from './data-context/setData'
-export { default as getData } from './data-context/getData'
-export { default as clearData } from './data-context/clearData'
-export { default as useValidation } from './data-context/useValidation'
-export { default as useSubmit } from './data-context/useSubmit'
+export { default as useData } from './DataContext/useData'
+export { default as setData } from './DataContext/setData'
+export { default as getData } from './DataContext/getData'
+export { default as clearData } from './DataContext/clearData'
+export { default as useValidation } from './DataContext/useValidation'
+export { default as useSubmit } from './DataContext/useSubmit'
 export { default as useTranslation } from '../hooks/useTranslation'
 export { default as useSnapshot } from '../hooks/useSnapshot'
